+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 - Fix "No space left" error on Git workspace [#1112](https://github.com/sider/runners/pull/1112)
 - **GolangCI-Lint** Explicitly set `--timeout` [#1117](https://github.com/sider/runners/pull/1117)
 - **Cppcheck** Add `addon` option [#1119](https://github.com/sider/runners/pull/1119)
+- **Cppcheck** 1.90 -> 2.0 [#1083](https://github.com/sider/runners/pull/1083)
 
 ## 0.24.0