+++ CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
-[Full diff](https://github.com/sider/runners/compare/0.24.0...HEAD)
+[Full diff](https://github.com/sider/runners/compare/0.25.0...HEAD)
+
+## 0.25.0
+
+[Full diff](https://github.com/sider/runners/compare/0.24.0...0.25.0)
 
 - Fix "No space left" error on Git workspace [#1112](https://github.com/sider/runners/pull/1112)
 - **GolangCI-Lint** Explicitly set `--timeout` [#1117](https://github.com/sider/runners/pull/1117)